+++ app/src/main/java/eco/org/greenapp/eco/org/greenapp/fragments/FragmentMyReviews.java
@@ -145,7 +145,7 @@ public class GetReviews extends AsyncTask<String, Void, String>{
 
                 review.setData_adaugare(adItem.getString("data"));
                 review.setContinut(adItem.getString("detalii"));
-                review.setNota(Integer.parseInt(adItem.getString("nota")));
+                review.setNota(Float.parseFloat(adItem.getString("nota")));
                 review.setUser(adItem.getString("username"));
                 review.setUrlProfil(adItem.getString("fotografie"));
                 lista.add(review);
